refactor: change from double to int in amount

+++ src/main/java/cl/transbank/webpay/example/controllers/WebpayPlusController.java
@@ -21,43 +21,46 @@ public class WebpayPlusController extends BaseController {
 
     private WebpayPlus.Transaction tx;
 
-    public WebpayPlusController(){
-        tx = new WebpayPlus.Transaction(new WebpayOptions(IntegrationCommerceCodes.WEBPAY_PLUS, IntegrationApiKeys.WEBPAY, IntegrationType.TEST));
+    public WebpayPlusController() {
+        tx = new WebpayPlus.Transaction(new WebpayOptions(IntegrationCommerceCodes.WEBPAY_PLUS,
+                IntegrationApiKeys.WEBPAY, IntegrationType.TEST));
     }
 
     @GetMapping("/create")
     public String createPage(HttpServletRequest req, Model model) {
-        Map<String, String> navigation = new LinkedHashMap<>() {{
-            put("request", "Petición");
-            put("response", "Respuesta");
-            put("form", "Formulario");
-        }};
+        Map<String, String> navigation = new LinkedHashMap<>() {
+            {
+                put("request", "Petición");
+                put("response", "Respuesta");
+                put("form", "Formulario");
+            }
+        };
         model.addAttribute("navigation", navigation);
 
-        Map<String, String> breadcrumbs = new LinkedHashMap<>() {{
-            put("Inicio", "/");
-            put("Webpay Plus", "/webpay-plus/");
-        }};
+        Map<String, String> breadcrumbs = new LinkedHashMap<>() {
+            {
+                put("Inicio", "/");
+                put("Webpay Plus", "/webpay-plus/");
+            }
+        };
         model.addAttribute("breadcrumbs", breadcrumbs);
 
         String buyOrder = "buyOrder_" + getRandomNumber();
         String sessionId = "sessionId_" + getRandomNumber();
-        double amount = 1000;
-        String returnUrl = req.getRequestURL().toString().replace("create","commit");
+        int amount = 1000;
+        String returnUrl = req.getRequestURL().toString().replace("create", "commit");
 
         Map<String, Object> request = Map.of(
                 "buyOrder", buyOrder,
                 "sessionId", sessionId,
                 "amount", amount,
-                "returnUrl", returnUrl
-        );
+                "returnUrl", returnUrl);
         model.addAttribute("request", request);
         try {
             var resp = tx.create(buyOrder, sessionId, amount, returnUrl);
             model.addAttribute("response_data", resp);
             model.addAttribute("response_data_json", toJson(resp));
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             log.error("ERROR", e);
         }
         return "webpay_plus/create";
